added name property to materials
When a keydown event is emitted and the "C" key is pressed, we traverse the object and check to see if the indexed child is a mesh and if its material name is "Book_Cover". But the materials that we've assigned to the colors array didn't have a name so the code within the condition wasn't executed as expected.